Refactor authentication and internationalization configuration: move config files to a dedicated directory and update import paths
