+++ middleware.ts
@@ -11,7 +11,7 @@ import {
   publicRoutes,
 } from "@/route";
 import NextAuth from "next-auth";
-import authConfig from "./auth.config";
+import authConfig from "./config/auth.config";
 
 const { auth } = NextAuth(authConfig);
